app/requirements.txt
beautifulsoup4
google-api-python-client 
google-auth-httplib2 
google-auth-oauthlib
requests
python-dateutil
python-dotenv
pytz
ciso8601